am263x: cpdma: Reset Recovery for CPSW
Reset recovery process can be triggered when CPSW or any mac port goes to hung
state. The usage is left for application specific usecases.
Currently we have 1 API EnetApp_triggerReset() which will take the callback
function as an argument to trigger reset for that specific SOC.

This is currently implemented in AM263 for following examples:
1. Layer2 CPSW switch
2. LwIP

Fixes: MCUSDK-12118

Signed-off-by: Surbhi Kapoor <s-kapoor1@ti.com>