all socs: CPSW: Fix the Genf ppm dir
In CPTS, the time stamp PPM adjustments depend on 'ts_ppm_dir' bit,
the time stamp is "increased" by PPM value when 'ts_ppm_dir' is cleared and vice versa.
The Genfn PPM adjustments depend on 'ts_genfN_ppm_dir' bit,
But, the genfn frequency is "decreased" by PPM value when 'ts_genfN_ppm_dir' bit is cleared and vice versa.
This is because the counters are adjusted with ppmvalue and not the frequecny directly.
As the Genfn ppm dir HW bit usage is different from time stamp ppm, add a new ENUM for Genfn PPM dir.

Fixes: MCUSDK-11390

Signed-off-by: Susheel <v-voora1@ti.com>