am64x/am243x : Enet : modify the sysconfig template for CPSW and ICSSG
-Enet module layer, Enet peripheral layer(partial) and Enet layer(partial)
are removed to make the code functionally safe. the purpose of removing
the layers are to remove function pointers, void pointers, type casting
and abstraction. the sysconfig templates for the CPSW and ICSSG should be
modified in order to incorporate those changes

The following changes aims to improve the overall code quality and meet
functional safety standards by reducing the use of potentially error-prone
constructs,

-removed enet module object in peripheral's module object
-removed enet peripheral object in CPSW and ICSSG object
-replaced Enet_Obj * with integer type
-encoded enet type and instId in hEnet

Signed-off-by: Yogeshwaran s <y-s1@ti.com>