am62xx: syscfg: Fix Core identification conditions
The Device Manager R5F core is referred to as "wkup-r5f"
for AM62Px and "r5f" or "dm-r5f" for AM62Dx and AM62Ax.

This causes regex pattern "wkup-r5f" to not match with DM-R5F
core as expected, instead "/r5f*/" is matched to both MCU-R5F
and DM-R5F on AM62Dx/AM62Ax platforms.

- Fix conditional statements to properly handle mcu-r5f and
  r5f/wkup-r5f segregation

Fixes: SITSW-7851

Signed-off-by: Nitika Verma <nitika.verma@ti.com>